Ingredients: placed all ingredient-related classes and exceptions in separate package
