+++ src/main/java/seedu/address/model/ingredient/Ingredient.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.ingredient;
 
 import java.util.Objects;
