+++ src/main/java/seedu/address/model/ingredient/IngredientName.java
@@ -1,4 +1,6 @@
-package seedu.address.model.person;
+package seedu.address.model.ingredient;
+
+import seedu.address.model.person.Name;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
