+++ src/main/java/seedu/address/model/ingredient/exceptions/IngredientNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.ingredient.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified ingredient.
